refactor: network still:((

+++ source/core/Defines.hpp
@@ -8,8 +8,8 @@
 #include <SFML/Graphics.hpp>
 
 
-const float WIDTH = 1280;
-const float HEIGHT = 720;
+const int WIDTH = 1280;
+const int HEIGHT = 720;
 
 const sf::Color HINT_COLOR1 = {0, 166, 147};
 const sf::Color HINT_COLOR2 = {100, 100, 100};
